requirements.txt
decord>=0.6.0
fvcore
matplotlib
mmcv-full==1.7.0
mmdet==2.25.1
mmpose==0.29.0
moviepy
numpy==1.24.3
opencv-contrib-python
opencv-python
pymemcache
scipy
torch>=1.5
tqdm
